+++ reservation-service/src/main/java/com/stos/concert/reservation/infrastructure/entity/Seat.java
@@ -1,4 +1,4 @@
-package com.stos.concert.reservation;
+package com.stos.concert.reservation.infrastructure.entity;
 
 import jakarta.persistence.Embeddable;
