+++ reservation-service/src/main/java/com/stos/concert/reservation/infrastructure/entity/TicketReservation.java
@@ -1,12 +1,10 @@
-package com.stos.concert.reservation;
+package com.stos.concert.reservation.infrastructure.entity;
 
-import jakarta.persistence.CollectionTable;
 import jakarta.persistence.ElementCollection;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
 import java.time.LocalDateTime;
 import java.util.List;
 import lombok.AllArgsConstructor;
