+++ reservation-service/src/main/java/com/stos/concert/reservation/infrastructure/repository/TicketReservationRepository.java
@@ -1,5 +1,6 @@
-package com.stos.concert.reservation;
+package com.stos.concert.reservation.infrastructure.repository;
 
+import com.stos.concert.reservation.infrastructure.entity.TicketReservation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TicketReservationRepository extends JpaRepository<TicketReservation, Long> {
